Update lint_and_security.yml
Key Corrections:

Indentation: The indentation has been corrected to follow the proper YAML structure for GitHub Actions workflows. Each nested level is indented by two spaces.

-fail_on_error Removed: The -fail_on_error flag has been temporarily removed from all reviewdog commands to help with troubleshooting the BrokenPipeError.

Explanation of YAML Structure:

name: The name of your workflow.

on: Specifies the events that trigger the workflow (push and pull requests to the main branch in this case).

jobs: Defines the jobs that will be executed as part of the workflow.

lint-and-security: The name of the job (you can choose any name you like).

runs-on: ubuntu-latest: Specifies the runner environment (Ubuntu in this case).

steps: A sequence of steps that will be executed within the job.

uses: Specifies a GitHub Action to use.

name: A descriptive name for the step.

env: Sets environment variables for the step.

run: Specifies a command to run.

with: Provides input parameters to an action.